Silver/requirements.txt
# Python dependencies for Silver Tier
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
watchdog>=3.0.0
python-dotenv>=1.0.0
